solve config problems
